Updated design and functionality
Added some new UI elements to the information center and portfolio. I also built a customer support system. The interface is a ticket style component that pops up whenever clicking a button. The system is only implemented on a few pages for this commit until I figure out an optimal UX for users to interact with the system.